+++ mysql/data_source_tables_test.go
@@ -17,9 +17,9 @@ func TestAccDataSourceTables(t *testing.T) {
 			{
 				Config: testAccTablesConfig_basic("mysql", "%"),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("data.mysql_tables.test", "database", "mysql"),
-					resource.TestCheckResourceAttr("data.mysql_tables.test", "pattern", "%"),
-					testAccTablesCount("data.mysql_tables.test", "tables.#", func(rn string, table_count int) error {
+					resource.TestCheckResourceAttr("data.singlestore_tables.test", "database", "mysql"),
+					resource.TestCheckResourceAttr("data.singlestore_tables.test", "pattern", "%"),
+					testAccTablesCount("data.singlestore_tables.test", "tables.#", func(rn string, table_count int) error {
 						if table_count < 1 {
 							return fmt.Errorf("%s: tables not found", rn)
 						}
@@ -31,9 +31,9 @@ func TestAccDataSourceTables(t *testing.T) {
 			{
 				Config: testAccTablesConfig_basic("mysql", "__table_does_not_exist__"),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("data.mysql_tables.test", "database", "mysql"),
-					resource.TestCheckResourceAttr("data.mysql_tables.test", "pattern", "__table_does_not_exist__"),
-					testAccTablesCount("data.mysql_tables.test", "tables.#", func(rn string, table_count int) error {
+					resource.TestCheckResourceAttr("data.singlestore_tables.test", "database", "mysql"),
+					resource.TestCheckResourceAttr("data.singlestore_tables.test", "pattern", "__table_does_not_exist__"),
+					testAccTablesCount("data.singlestore_tables.test", "tables.#", func(rn string, table_count int) error {
 						if table_count > 0 {
 							return fmt.Errorf("%s: unexpected table found", rn)
 						}
@@ -72,7 +72,7 @@ func testAccTablesCount(rn string, key string, check func(string, int) error) re
 
 func testAccTablesConfig_basic(database string, pattern string) string {
 	return fmt.Sprintf(`
-data "mysql_tables" "test" {
+data "singlestore_tables" "test" {
 		database = "%s"
 		pattern = "%s"
 }`, database, pattern)