+++ mysql/resource_database.go
@@ -4,15 +4,16 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
+	"strconv"
 	"strings"
 
-	"github.com/go-sql-driver/mysql"
-	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
-const defaultCharacterSetKeyword = "CHARACTER SET "
-const defaultCollateKeyword = "COLLATE "
+const defaultPartitionsKeyword = "PARTITIONS "
+const defaultS3PathKeyword = "ON S3 "
+const defaultConfigKeyword = "CONFIG "
+const defaultCredentialsKeyword = "CREDENTIALS '{}'"
 const unknownDatabaseErrCode = 1049
 
 func resourceDatabase() *schema.Resource {
@@ -31,16 +32,22 @@ func resourceDatabase() *schema.Resource {
 				ForceNew: true,
 			},
 
-			"default_character_set": {
-				Type:     schema.TypeString,
+			"partitions": {
+				Type:     schema.TypeInt,
 				Optional: true,
-				Default:  "utf8",
+				Default:  0,
 			},
 
-			"default_collation": {
+			"s3_path": {
 				Type:     schema.TypeString,
 				Optional: true,
-				Default:  "utf8_general_ci",
+				Default:  "",
+			},
+
+			"config": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Default:  "{\"region\":\"us-east-1\"}",
 			},
 		},
 	}
@@ -89,67 +96,25 @@ func ReadDatabase(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	// This is kinda flimsy-feeling, since it depends on the formatting
-	// of the SHOW CREATE DATABASE output... but this data doesn't seem
+	// of the SHOW DATABASES output... but this data doesn't seem
 	// to be available any other way, so hopefully MySQL keeps this
 	// compatible in future releases.
 
 	name := d.Id()
-	stmtSQL := "SHOW CREATE DATABASE " + quoteIdentifier(name)
+	stmtSQL := "SHOW DATABASES LIKE " + quoteIdentifier(name)
 
 	log.Println("Executing statement:", stmtSQL)
-	var createSQL, _database string
-	err = db.QueryRow(stmtSQL).Scan(&_database, &createSQL)
+	var _database string
+	err = db.QueryRow(stmtSQL).Scan(&_database)
 	if err != nil {
-		if mysqlErr, ok := err.(*mysql.MySQLError); ok {
-			if mysqlErr.Number == unknownDatabaseErrCode {
-				d.SetId("")
-				return nil
-			}
+		if err == sql.ErrNoRows {
+			d.SetId("")
+			return nil
 		}
-		return fmt.Errorf("Error during show create database: %s", err)
-	}
-
-	defaultCharset := extractIdentAfter(createSQL, defaultCharacterSetKeyword)
-	defaultCollation := extractIdentAfter(createSQL, defaultCollateKeyword)
-
-	if defaultCollation == "" && defaultCharset != "" {
-		// MySQL doesn't return the collation if it's the default one for
-		// the charset, so if we don't have a collation we need to go
-		// hunt for the default.
-		stmtSQL := "SHOW COLLATION WHERE `Charset` = ? AND `Default` = 'Yes'"
-		var empty interface{}
-
-		requiredVersion, _ := version.NewVersion("8.0.0")
-		currentVersion, err := serverVersion(db)
-		if err != nil {
-			return err
-		}
-
-		serverVersionString, err := serverVersionString(db)
-		if err != nil {
-			return err
-		}
-
-		// MySQL 8 returns more data in a row.
-		var res error
-		if !strings.Contains(serverVersionString, "MariaDB") && currentVersion.GreaterThan(requiredVersion) {
-			res = db.QueryRow(stmtSQL, defaultCharset).Scan(&defaultCollation, &empty, &empty, &empty, &empty, &empty, &empty)
-		} else {
-			res = db.QueryRow(stmtSQL, defaultCharset).Scan(&defaultCollation, &empty, &empty, &empty, &empty, &empty)
-		}
-
-		if res != nil {
-			if res == sql.ErrNoRows {
-				return fmt.Errorf("Charset %s has no default collation", defaultCharset)
-			}
-
-			return fmt.Errorf("Error getting default charset: %s, %s", res, defaultCharset)
-		}
+		return fmt.Errorf("Error during show databases: %s", err)
 	}
 
 	d.Set("name", name)
-	d.Set("default_character_set", defaultCharset)
-	d.Set("default_collation", defaultCollation)
 
 	return nil
 }
@@ -173,25 +138,33 @@ func DeleteDatabase(d *schema.ResourceData, meta interface{}) error {
 
 func databaseConfigSQL(verb string, d *schema.ResourceData) string {
 	name := d.Get("name").(string)
-	defaultCharset := d.Get("default_character_set").(string)
-	defaultCollation := d.Get("default_collation").(string)
+	defaultPartitions := d.Get("partitions").(int)
+	defaultS3Path := d.Get("s3_path").(string)
+	defaultConfig := d.Get("config").(string)
 
-	var defaultCharsetClause string
-	var defaultCollationClause string
+	var defaultPartitionsClause string
+	var defaultS3PathClause string
+	var defaultConfigClause string
+	var bottomlessStatement string
 
-	if defaultCharset != "" {
-		defaultCharsetClause = defaultCharacterSetKeyword + quoteIdentifier(defaultCharset)
+	if defaultPartitions != 0 {
+		defaultPartitionsClause = defaultPartitionsKeyword + strconv.Itoa(defaultPartitions)
 	}
-	if defaultCollation != "" {
-		defaultCollationClause = defaultCollateKeyword + quoteIdentifier(defaultCollation)
+	if defaultS3Path != "" {
+		// bottomless database
+		defaultS3PathClause = defaultS3PathKeyword + quoteIdentifier(defaultS3Path)
+		if defaultConfig != "" {
+			defaultConfigClause = defaultConfigKeyword + quote(defaultConfig)
+		}
+		bottomlessStatement = fmt.Sprintf("%s %s %s", defaultS3PathClause, defaultConfigClause, defaultCredentialsKeyword)
 	}
 
 	return fmt.Sprintf(
 		"%s DATABASE %s %s %s",
 		verb,
 		quoteIdentifier(name),
-		defaultCharsetClause,
-		defaultCollationClause,
+		defaultPartitionsClause,
+		bottomlessStatement,
 	)
 }