+++ mysql/resource_database_test.go
@@ -2,7 +2,6 @@ package mysql
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/go-sql-driver/mysql"
@@ -22,7 +21,7 @@ func TestAccDatabase(t *testing.T) {
 			{
 				Config: testAccDatabaseConfig_basic(dbName),
 				Check: testAccDatabaseCheck_basic(
-					"mysql_database.test", dbName,
+					"singlestore_database.test", dbName,
 				),
 			},
 		},
@@ -32,12 +31,7 @@ func TestAccDatabase(t *testing.T) {
 func TestAccDatabase_collationChange(t *testing.T) {
 	dbName := "terraform_acceptance_test"
 
-	charset1 := "latin1"
-	charset2 := "utf8"
-	collation1 := "latin1_bin"
-	collation2 := "utf8_general_ci"
-
-	resourceName := "mysql_database.test"
+	resourceName := "singlestore_database.test"
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -45,9 +39,9 @@ func TestAccDatabase_collationChange(t *testing.T) {
 		CheckDestroy: testAccDatabaseCheckDestroy(dbName),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDatabaseConfig_full(dbName, charset1, collation1),
+				Config: testAccDatabaseConfig_full(dbName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccDatabaseCheck_full("mysql_database.test", dbName, charset1, collation1),
+					testAccDatabaseCheck_full("singlestore_database.test", dbName),
 				),
 			},
 			{
@@ -56,17 +50,9 @@ func TestAccDatabase_collationChange(t *testing.T) {
 				ImportStateVerify: true,
 			},
 			{
-				PreConfig: func() {
-					db, err := connectToMySQL(testAccProvider.Meta().(*MySQLConfiguration))
-					if err != nil {
-						return
-					}
-
-					db.Exec(fmt.Sprintf("ALTER DATABASE %s CHARACTER SET %s COLLATE %s", dbName, charset2, collation2))
-				},
-				Config: testAccDatabaseConfig_full(dbName, charset1, collation1),
+				Config: testAccDatabaseConfig_full(dbName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccDatabaseCheck_full(resourceName, dbName, charset1, collation1),
+					testAccDatabaseCheck_full(resourceName, dbName),
 				),
 			},
 		},
@@ -74,10 +60,10 @@ func TestAccDatabase_collationChange(t *testing.T) {
 }
 
 func testAccDatabaseCheck_basic(rn string, name string) resource.TestCheckFunc {
-	return testAccDatabaseCheck_full(rn, name, "utf8", "utf8_bin")
+	return testAccDatabaseCheck_full(rn, name)
 }
 
-func testAccDatabaseCheck_full(rn string, name string, charset string, collation string) resource.TestCheckFunc {
+func testAccDatabaseCheck_full(rn string, name string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[rn]
 		if !ok {
@@ -97,13 +83,6 @@ func testAccDatabaseCheck_full(rn string, name string, charset string, collation
 		err = db.QueryRow(fmt.Sprintf("SHOW CREATE DATABASE %s", name)).Scan(&_name, &createSQL)
 		if err != nil {
 			return fmt.Errorf("error reading database: %s", err)
-		}
-
-		if strings.Index(createSQL, fmt.Sprintf("CHARACTER SET %s", charset)) == -1 {
-			return fmt.Errorf("database default charset isn't %s", charset)
-		}
-		if strings.Index(createSQL, fmt.Sprintf("COLLATE %s", collation)) == -1 {
-			return fmt.Errorf("database default collation isn't %s", collation)
 		}
 
 		return nil
@@ -134,14 +113,12 @@ func testAccDatabaseCheckDestroy(name string) resource.TestCheckFunc {
 }
 
 func testAccDatabaseConfig_basic(name string) string {
-	return testAccDatabaseConfig_full(name, "utf8", "utf8_bin")
+	return testAccDatabaseConfig_full(name)
 }
 
-func testAccDatabaseConfig_full(name string, charset string, collation string) string {
+func testAccDatabaseConfig_full(name string) string {
 	return fmt.Sprintf(`
-resource "mysql_database" "test" {
+resource "singlestore_database" "test" {
     name = "%s"
-    default_character_set = "%s"
-    default_collation = "%s"
-}`, name, charset, collation)
+}`, name)
 }