+++ mysql/resource_sql.go
@@ -1,6 +1,7 @@
 package mysql
 
 import (
+	"fmt"
 	"log"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -10,10 +11,16 @@ func resourceSql() *schema.Resource {
 	return &schema.Resource{
 		Create: CreateSql,
 		Read:   ReadSql,
+		Update: UpdateSql,
 		Delete: DeleteSql,
 
 		Schema: map[string]*schema.Schema{
 			"name": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"database_name": {
 				Type:     schema.TypeString,
 				Required: true,
 				ForceNew: true,
@@ -23,10 +30,15 @@ func resourceSql() *schema.Resource {
 				Required: true,
 				ForceNew: true,
 			},
+			"update_sql": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Default:  "select 1",
+			},
 			"delete_sql": {
 				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
+				Optional: true,
+				Default:  "select 1",
 			},
 		},
 	}
@@ -38,11 +50,34 @@ func CreateSql(d *schema.ResourceData, meta interface{}) error {
 		return err
 	}
 	name := d.Get("name").(string)
-	create_sql := d.Get("create_sql").(string)
+	databaseName := d.Get("database_name").(string)
+	createSql := fmt.Sprintf("BEGIN; USE %s; %s COMMIT;", databaseName, d.Get("create_sql").(string))
 
-	log.Println("Executing SQL", create_sql)
+	log.Println("Executing SQL", createSql)
 
-	_, err = db.Exec(create_sql)
+	_, err = db.Exec(createSql)
+
+	if err != nil {
+		return err
+	}
+
+	d.SetId(name)
+
+	return nil
+}
+
+func UpdateSql(d *schema.ResourceData, meta interface{}) error {
+	db, err := meta.(*MySQLConfiguration).GetDbConn()
+	if err != nil {
+		return err
+	}
+	name := d.Get("name").(string)
+	databaseName := d.Get("database_name").(string)
+	updateSql := fmt.Sprintf("BEGIN; USE %s; %s COMMIT;", databaseName, d.Get("update_sql").(string))
+
+	log.Println("Executing SQL", updateSql)
+
+	_, err = db.Exec(updateSql)
 
 	if err != nil {
 		return err
@@ -62,7 +97,8 @@ func DeleteSql(d *schema.ResourceData, meta interface{}) error {
 	if err != nil {
 		return err
 	}
-	delete_sql := d.Get("delete_sql").(string)
+	databaseName := d.Get("database_name").(string)
+	delete_sql := fmt.Sprintf("BEGIN; USE %s; %s COMMIT;", databaseName, d.Get("delete_sql").(string))
 
 	log.Println("Executing SQL:", delete_sql)